+++ src/lib/fragments/AgGrid.react.js
@@ -42,13 +42,12 @@ import RowMenuRenderer from '../renderers/rowMenuRenderer';
 import {customFunctions} from '../renderers/customFunctions';
 
 import {AgGridReact, useGridFilter} from 'ag-grid-react';
-import {
-    themeAlpine,
-    themeBalham,
-    themeMaterial,
-    themeQuartz,
-} from 'ag-grid-community';
-const themes = {themeAlpine, themeBalham, themeMaterial, themeQuartz};
+import * as agGrid from 'ag-grid-community';
+let themes = {};
+if (agGrid && typeof agGrid === 'object' && !Array.isArray(agGrid)) {
+    const {themeAlpine, themeBalham, themeMaterial, themeQuartz} = agGrid;
+    themes = {themeAlpine, themeBalham, themeMaterial, themeQuartz};
+}
 
 // d3 imports
 import * as d3Format from 'd3-format';
@@ -157,6 +156,7 @@ export function DashAgGrid(props) {
                     esprima.parse(funcString).body[0].expression;
                 const context = {
                     ...themes,
+                    agGrid,
                     d3,
                     dash_clientside,
                     ...customFunctions,
@@ -175,6 +175,7 @@ export function DashAgGrid(props) {
                     esprima.parse(funcString).body[0].expression;
                 const context = {
                     ...themes,
+                    agGrid,
                     d3,
                     dash_clientside,
                     ...customFunctions,
@@ -195,6 +196,7 @@ export function DashAgGrid(props) {
                     esprima.parse(funcString).body[0].expression;
                 const context = {
                     ...themes,
+                    agGrid,
                     d3,
                     ...customFunctions,
                     ...window.dashAgGridFunctions,